fix(metrics): address review findings on the canonical metric pipeline
Correctness:
- Reject non-finite doubles instead of accepting a point whose value is
  silently stored as NULL, and reject non-finite histogram sum/min/max.
- Render the canonical payload from the same canonical view the columns
  store, so PointId can never disagree with the persisted row. This also
  fixes summaryQuantilesJson storing raw input, and `{asInt: null,
  asDouble: 5}` being persisted as int 0.
- Order canonical attributes by code unit, not localeCompare: identity
  hashing must not depend on the host locale or ICU build.
- Bound timestamps by the ECMA-262 Date maximum, reachable via an
  exemplar's unvalidated nanos.
- Validate OTLP containers before iterating, so a malformed resourceMetrics
  is a counted rejection rather than a TypeError past the per-point catch.
- Derive affected rollup buckets from sequence dependency, not temporality:
  summaries are cumulative while reporting no temporality, so a late one
  left the next bucket stale.
- Normalize exponential-histogram zero thresholds per the OTel data model
  (adopt the largest, widen off a bisected bucket, fold covered buckets
  into zeroCount) instead of unioning incompatible layouts and writing null.
- Run canonical validation whenever a request carries metrics, so a batch
  whose dataPoints are all malformed is no longer acked as fully accepted.
- Replace reserved langwatch.origin.*/ingestion_source.* keys rather than
  appending duplicates, on traces and logs as well as metrics.
- Tighten the correlation event schema to the command's constraints; a
  replayed event is re-validated, so the looser schema let malformed rows
  back into the folds.

Security/privacy:
- Return a stable message for persistence failures; queue and ClickHouse
  errors no longer reach the OTLP response.
- Mark the organization.id span attribute as intentional exposure.
- Tenant-scope the metric correlation job/idempotency key, per convention.

Performance:
- Fetch each affected rollup bucket as its own range instead of one
  min..max span across a potentially distant next sample.
- Give bulkAppend real bulk repository operations: replay chunks now cost
  one round trip per table, and rollups are grouped per series.

Maintainability:
- Split canonicalMetric.ts (947 lines) and rollup.ts (537) into focused
  modules under 300; split the repository's raw/usage responsibilities.
- Named object parameters across the new APIs; split the oversized test
  files by concern.
- Rename metric_record_count to metric_exemplar_correlation_count: the
  folds see exemplar correlations, not data points.
- Document the organization-scoped billing-ledger carve-out to the
  TenantId-first rule, with a test pinning it.
- Add the irreversible-rollback marker to migration 00042 and correct the
  storage-meter docs to the filtered table set.
