+++ dev/docs/best_practices/clickhouse-queries.md
@@ -170,6 +170,22 @@ Current sweeps under this carve-out, both on worker boot:
 
 Anything else that wants to skip the filter should be a repository method taking a `tenantId`, not a sweep.
 
+### Carve-out: organization-scoped billing ledgers
+
+A table whose whole purpose is to total usage *across* an organization's projects cannot lead with `TenantId` — the aggregate it exists to answer has no single tenant. One table qualifies today:
+
+- `metric_usage_estimates` (`queryMetricUsageEstimates`, `metric-data-point.usage.ts`) — ORDER BY `(OrganizationId, TenantId, PointId)`.
+
+It is allowed to lead with `OrganizationId` **only** because all of these hold:
+
+1. The ClickHouse instance is itself selected by organization (`getClickHouseClientForOrganization`, and `getClickHouseClientForProject` resolves project → org → client). An organization's query physically cannot reach another organization's rows, so `OrganizationId` is the same isolation boundary `TenantId` is elsewhere — not a weaker one.
+2. `OrganizationId` leads the table's sort key, so the predicate order matches the index. A `TenantId`-first predicate would be both wrong for the aggregate and worse for the scan.
+3. `TenantId` is still ANDed in whenever the caller supplies one, and remains a selected grouping dimension.
+4. The table holds identifiers and byte counts only — never attributes, values, buckets or payloads — so a scoping mistake cannot leak customer data.
+5. A test pins that the organization-wide path uses the organization-resolved client and filters on `OrganizationId`.
+
+A new table wanting this carve-out needs all five, plus a line here. Anything that merely *finds it convenient* to skip `TenantId` does not qualify.
+
 ## Code Review Checklist
 
 When reviewing a PR that touches a `*.clickhouse.repository.ts` or any service hitting ClickHouse, scan for:
