+++ langwatch/src/server/app-layer/metrics/repositories/metric-data-point.clickhouse.repository.ts
@@ -7,198 +7,34 @@ import {
 } from "~/server/event-sourcing/pipelines/metric-processing/rollup";
 import { METRIC_ROLLUP_INTERVAL_MS } from "~/server/event-sourcing/pipelines/metric-processing/schemas/constants";
 import type {
-  AggregationTemporality,
   CanonicalMetricDataPoint,
-  MetricKind,
+  MetricRollupRow,
   MetricUsageEstimate,
   MetricUsageEstimateQuery,
 } from "~/server/event-sourcing/pipelines/metric-processing/schemas/metricDataPoint";
 import { SecurityError } from "~/server/event-sourcing/services/errorHandling";
-import { EventUtils } from "~/server/event-sourcing/utils/event.utils";
-import type { MetricDataPointRepository } from "./metric-data-point.repository";
+import type {
+  MetricDataPointBulkWrite,
+  MetricDataPointRepository,
+  MetricDataPointWrite,
+} from "./metric-data-point.repository";
+import {
+  fromRaw,
+  RAW_SELECT,
+  rawRow,
+  rollupRow,
+  seriesRow,
+  usageEstimateRow,
+  validatePoint,
+  type RawMetricRow,
+} from "./metric-data-point.rows";
+import { queryMetricUsageEstimates } from "./metric-data-point.usage";
 
 const logger = createLogger(
   "langwatch:app-layer:metrics:metric-data-point-repository",
 );
 
-interface RawMetricRow {
-  TenantId: string;
-  PointId: string;
-  SeriesId: string;
-  ResourceSchemaUrl: string;
-  ResourceAttributesJson: string;
-  ResourceAttributeKeys: string[];
-  ScopeSchemaUrl: string;
-  ScopeName: string;
-  ScopeVersion: string;
-  ScopeAttributesJson: string;
-  ScopeAttributeKeys: string[];
-  MetricName: string;
-  MetricDescription: string;
-  MetricUnit: string;
-  MetricKind: MetricKind;
-  AggregationTemporality: AggregationTemporality;
-  IsMonotonic: number | boolean | null;
-  PointAttributesJson: string;
-  PointAttributeKeys: string[];
-  StartTimeUnixNano: string;
-  TimeUnixNano: string;
-  TimeUnixMs: string | number;
-  Flags: number;
-  ValueType: "none" | "int" | "double";
-  ValueInt: string | null;
-  ValueDouble: number | null;
-  Count: string | null;
-  Sum: number | null;
-  Min: number | null;
-  Max: number | null;
-  ExplicitBounds: number[];
-  BucketCounts: string[];
-  ExponentialScale: number | null;
-  ExponentialZeroThreshold: number | null;
-  ZeroCount: string | null;
-  PositiveOffset: number | null;
-  PositiveBucketCounts: string[];
-  NegativeOffset: number | null;
-  NegativeBucketCounts: string[];
-  SummaryQuantilesJson: string;
-  CanonicalPayload: string;
-  _size_bytes: number;
-  OccurredAt: string | number;
-  AcceptedAt: string | number;
-}
-
-const RAW_SELECT = `
-  TenantId, PointId, SeriesId,
-  ResourceSchemaUrl, ResourceAttributesJson, ResourceAttributeKeys,
-  ScopeSchemaUrl, ScopeName, ScopeVersion, ScopeAttributesJson, ScopeAttributeKeys,
-  MetricName, MetricDescription, MetricUnit, MetricKind,
-  AggregationTemporality, IsMonotonic,
-  PointAttributesJson, PointAttributeKeys,
-  StartTimeUnixNano, TimeUnixNano, toUnixTimestamp64Milli(TimeUnixMs) AS TimeUnixMs,
-  Flags, ValueType, ValueInt, ValueDouble, Count, Sum, Min, Max,
-  ExplicitBounds, BucketCounts, ExponentialScale, ExponentialZeroThreshold, ZeroCount,
-  PositiveOffset, PositiveBucketCounts, NegativeOffset, NegativeBucketCounts,
-  SummaryQuantilesJson, CanonicalPayload, _size_bytes,
-  toUnixTimestamp64Milli(OccurredAt) AS OccurredAt,
-  toUnixTimestamp64Milli(AcceptedAt) AS AcceptedAt
-`;
-
-function rawRow(point: CanonicalMetricDataPoint, retentionDays: number) {
-  return {
-    TenantId: point.tenantId,
-    PointId: point.pointId,
-    SeriesId: point.seriesId,
-    ResourceSchemaUrl: point.resourceSchemaUrl,
-    ResourceAttributesJson: point.resourceAttributesJson,
-    ResourceAttributeKeys: point.resourceAttributeKeys,
-    ScopeSchemaUrl: point.scopeSchemaUrl,
-    ScopeName: point.scopeName,
-    ScopeVersion: point.scopeVersion,
-    ScopeAttributesJson: point.scopeAttributesJson,
-    ScopeAttributeKeys: point.scopeAttributeKeys,
-    MetricName: point.metricName,
-    MetricDescription: point.metricDescription,
-    MetricUnit: point.metricUnit,
-    MetricKind: point.metricKind,
-    AggregationTemporality: point.aggregationTemporality,
-    IsMonotonic: point.isMonotonic,
-    PointAttributesJson: point.pointAttributesJson,
-    PointAttributeKeys: point.pointAttributeKeys,
-    StartTimeUnixNano: point.startTimeUnixNano,
-    TimeUnixNano: point.timeUnixNano,
-    TimeUnixMs: new Date(point.timeUnixMs),
-    Flags: point.flags,
-    ValueType: point.valueType,
-    ValueInt: point.valueInt,
-    ValueDouble: point.valueDouble,
-    Count: point.count,
-    Sum: point.sum,
-    Min: point.min,
-    Max: point.max,
-    ExplicitBounds: point.explicitBounds,
-    BucketCounts: point.bucketCounts,
-    ExponentialScale: point.exponentialScale,
-    ExponentialZeroThreshold: point.exponentialZeroThreshold,
-    ZeroCount: point.zeroCount,
-    PositiveOffset: point.positiveOffset,
-    PositiveBucketCounts: point.positiveBucketCounts,
-    NegativeOffset: point.negativeOffset,
-    NegativeBucketCounts: point.negativeBucketCounts,
-    SummaryQuantilesJson: point.summaryQuantilesJson,
-    CanonicalPayload: point.canonicalPayload,
-    OccurredAt: new Date(point.occurredAt),
-    AcceptedAt: new Date(point.acceptedAt),
-    // Keep the first acceptance when the same PointId is retried.
-    DedupVersion: (
-      18_446_744_073_709_551_615n - BigInt(point.acceptedAt)
-    ).toString(),
-    _retention_days: retentionDays,
-    _size_bytes: point.canonicalSizeBytes,
-  };
-}
-
-function fromRaw(
-  row: RawMetricRow,
-  organizationId: string,
-): CanonicalMetricDataPoint {
-  return {
-    tenantId: row.TenantId,
-    // Organization identity is deliberately absent from authoritative metric
-    // storage. It is carried only long enough to write the shadow ledger.
-    organizationId,
-    pointId: row.PointId,
-    seriesId: row.SeriesId,
-    resourceSchemaUrl: row.ResourceSchemaUrl,
-    resourceAttributesJson: row.ResourceAttributesJson,
-    resourceAttributeKeys: row.ResourceAttributeKeys,
-    scopeSchemaUrl: row.ScopeSchemaUrl,
-    scopeName: row.ScopeName,
-    scopeVersion: row.ScopeVersion,
-    scopeAttributesJson: row.ScopeAttributesJson,
-    scopeAttributeKeys: row.ScopeAttributeKeys,
-    metricName: row.MetricName,
-    metricDescription: row.MetricDescription,
-    metricUnit: row.MetricUnit,
-    metricKind: row.MetricKind,
-    aggregationTemporality: row.AggregationTemporality,
-    isMonotonic: row.IsMonotonic === null ? null : Boolean(row.IsMonotonic),
-    pointAttributesJson: row.PointAttributesJson,
-    pointAttributeKeys: row.PointAttributeKeys,
-    startTimeUnixNano: String(row.StartTimeUnixNano),
-    timeUnixNano: String(row.TimeUnixNano),
-    timeUnixMs: Number(row.TimeUnixMs),
-    flags: row.Flags,
-    valueType: row.ValueType,
-    valueInt: row.ValueInt === null ? null : String(row.ValueInt),
-    valueDouble: row.ValueDouble,
-    count: row.Count === null ? null : String(row.Count),
-    sum: row.Sum,
-    min: row.Min,
-    max: row.Max,
-    explicitBounds: row.ExplicitBounds,
-    bucketCounts: row.BucketCounts.map(String),
-    exponentialScale: row.ExponentialScale,
-    exponentialZeroThreshold: row.ExponentialZeroThreshold,
-    zeroCount: row.ZeroCount === null ? null : String(row.ZeroCount),
-    positiveOffset: row.PositiveOffset,
-    positiveBucketCounts: row.PositiveBucketCounts.map(String),
-    negativeOffset: row.NegativeOffset,
-    negativeBucketCounts: row.NegativeBucketCounts.map(String),
-    summaryQuantilesJson: row.SummaryQuantilesJson,
-    canonicalPayload: row.CanonicalPayload,
-    canonicalSizeBytes: Number(row._size_bytes),
-    occurredAt: Number(row.OccurredAt),
-    acceptedAt: Number(row.AcceptedAt),
-  };
-}
-
-function validate(point: CanonicalMetricDataPoint, operation: string): void {
-  EventUtils.validateTenantId({ tenantId: point.tenantId }, operation);
-  if (!/^[a-f0-9]{64}$/.test(point.pointId)) {
-    throw new SecurityError(operation, "invalid PointId", point.tenantId);
-  }
-}
+const INSERT_SETTINGS = { async_insert: 1, wait_for_async_insert: 1 } as const;
 
 export class MetricDataPointClickHouseRepository implements MetricDataPointRepository {
   constructor(
@@ -206,89 +42,170 @@ export class MetricDataPointClickHouseRepository implements MetricDataPointRepos
     private readonly resolveOrganizationClient: ClickHouseClientResolver = resolveClient,
   ) {}
 
-  async ensureDataPoint(
-    point: CanonicalMetricDataPoint,
+  async ensureDataPoint(args: MetricDataPointWrite): Promise<void> {
+    await this.ensureDataPoints({
+      points: [args.point],
+      retentionDays: args.retentionDays,
+    });
+  }
+
+  /** One round trip per table, however many points the chunk holds. */
+  async ensureDataPoints({
+    points,
     retentionDays = PLATFORM_DEFAULT_RETENTION_DAYS,
-  ): Promise<void> {
-    validate(point, "MetricDataPointClickHouseRepository.ensureDataPoint");
-    const client = await this.resolveClient(point.tenantId);
+  }: MetricDataPointBulkWrite): Promise<void> {
+    if (points.length === 0) return;
+    for (const point of points) {
+      validatePoint({
+        point,
+        operation: "MetricDataPointClickHouseRepository.ensureDataPoints",
+      });
+    }
+    const client = await this.resolveClient(points[0]!.tenantId);
     try {
       // Raw must be authoritative before any derived or shadow write.
       await client.insert({
         table: "metric_data_points",
-        values: [rawRow(point, retentionDays)],
+        values: points.map((point) => rawRow({ point, retentionDays })),
         format: "JSONEachRow",
-        clickhouse_settings: { async_insert: 1, wait_for_async_insert: 1 },
+        clickhouse_settings: INSERT_SETTINGS,
       });
       await client.insert({
         table: "metric_usage_estimates",
-        values: [
-          {
-            OrganizationId: point.organizationId,
-            TenantId: point.tenantId,
-            PointId: point.pointId,
-            SeriesId: point.seriesId,
-            MetricName: point.metricName,
-            AcceptedAt: new Date(point.acceptedAt),
-            AcceptedHour: new Date(
-              Math.floor(point.acceptedAt / 3_600_000) * 3_600_000,
-            ),
-            CanonicalSourceBytes: point.canonicalSizeBytes,
-            // ReplacingMergeTree keeps the largest version. Inverting the
-            // acceptance millisecond makes the first accepted retry win.
-            DedupVersion: (
-              18_446_744_073_709_551_615n - BigInt(point.acceptedAt)
-            ).toString(),
-          },
-        ],
+        values: points.map(usageEstimateRow),
         format: "JSONEachRow",
-        clickhouse_settings: { async_insert: 1, wait_for_async_insert: 1 },
+        clickhouse_settings: INSERT_SETTINGS,
       });
     } catch (error) {
       logger.error(
-        { tenantId: point.tenantId, pointId: point.pointId, error },
-        "Failed to persist canonical metric point",
+        {
+          tenantId: points[0]!.tenantId,
+          pointCount: points.length,
+          error,
+        },
+        "Failed to persist canonical metric points",
       );
       throw error;
     }
   }
 
-  async upsertSeries(
-    point: CanonicalMetricDataPoint,
+  async upsertSeries(args: MetricDataPointWrite): Promise<void> {
+    await this.upsertSeriesMany({
+      points: [args.point],
+      retentionDays: args.retentionDays,
+    });
+  }
+
+  async upsertSeriesMany({
+    points,
     retentionDays = PLATFORM_DEFAULT_RETENTION_DAYS,
-  ): Promise<void> {
-    validate(point, "MetricDataPointClickHouseRepository.upsertSeries");
-    const client = await this.resolveClient(point.tenantId);
+  }: MetricDataPointBulkWrite): Promise<void> {
+    if (points.length === 0) return;
+    for (const point of points) {
+      validatePoint({
+        point,
+        operation: "MetricDataPointClickHouseRepository.upsertSeriesMany",
+      });
+    }
+    // LastSeenAt is the replacement version, so only the newest point per
+    // series can win. Collapsing here writes one row per series instead of one
+    // per point and leaves the merge with nothing to undo.
+    const latest = new Map<string, CanonicalMetricDataPoint>();
+    for (const point of points) {
+      const current = latest.get(point.seriesId);
+      if (!current || point.timeUnixMs > current.timeUnixMs) {
+        latest.set(point.seriesId, point);
+      }
+    }
+    const client = await this.resolveClient(points[0]!.tenantId);
     await client.insert({
       table: "metric_series",
-      values: [
-        {
-          TenantId: point.tenantId,
-          SeriesId: point.seriesId,
-          ResourceSchemaUrl: point.resourceSchemaUrl,
-          ResourceAttributesJson: point.resourceAttributesJson,
-          ResourceAttributeKeys: point.resourceAttributeKeys,
-          ScopeSchemaUrl: point.scopeSchemaUrl,
-          ScopeName: point.scopeName,
-          ScopeVersion: point.scopeVersion,
-          ScopeAttributesJson: point.scopeAttributesJson,
-          ScopeAttributeKeys: point.scopeAttributeKeys,
-          MetricName: point.metricName,
-          MetricDescription: point.metricDescription,
-          MetricUnit: point.metricUnit,
-          MetricKind: point.metricKind,
-          AggregationTemporality: point.aggregationTemporality,
-          IsMonotonic: point.isMonotonic,
-          PointAttributesJson: point.pointAttributesJson,
-          PointAttributeKeys: point.pointAttributeKeys,
-          LastSeenAt: new Date(point.timeUnixMs),
-          _retention_days: retentionDays,
-          _size_bytes: 0,
-        },
-      ],
+      values: [...latest.values()].map((point) =>
+        seriesRow({ point, retentionDays }),
+      ),
       format: "JSONEachRow",
-      clickhouse_settings: { async_insert: 1, wait_for_async_insert: 1 },
+      clickhouse_settings: INSERT_SETTINGS,
     });
+  }
+
+  async recomputeAffectedRollups(args: MetricDataPointWrite): Promise<void> {
+    await this.recomputeAffectedRollupsMany({
+      points: [args.point],
+      retentionDays: args.retentionDays,
+    });
+  }
+
+  /**
+   * Recomputes a chunk's rollups per series rather than per point: the
+   * authoritative fetch and the row write are what cost round trips, and every
+   * point of a series shares them. Ensuring the raw points up front also makes
+   * the result independent of the order the chunk happens to arrive in.
+   */
+  async recomputeAffectedRollupsMany({
+    points,
+    retentionDays = PLATFORM_DEFAULT_RETENTION_DAYS,
+  }: MetricDataPointBulkWrite): Promise<void> {
+    if (points.length === 0) return;
+    // Projection queues are independent. Ensuring the points here makes the
+    // raw-before-derived invariant true even if this projection wins the race.
+    await this.ensureDataPoints({ points, retentionDays });
+
+    const bySeries = new Map<string, CanonicalMetricDataPoint[]>();
+    for (const point of points) {
+      bySeries.set(point.seriesId, [
+        ...(bySeries.get(point.seriesId) ?? []),
+        point,
+      ]);
+    }
+
+    const rows: MetricRollupRow[] = [];
+    for (const seriesPoints of bySeries.values()) {
+      rows.push(...(await this.rollupRowsForSeries(seriesPoints)));
+    }
+    if (rows.length === 0) return;
+
+    const client = await this.resolveClient(points[0]!.tenantId);
+    await client.insert({
+      table: "metric_time_rollups",
+      values: rows.map((row) => rollupRow({ row, retentionDays })),
+      format: "JSONEachRow",
+      clickhouse_settings: INSERT_SETTINGS,
+    });
+  }
+
+  private async rollupRowsForSeries(
+    points: CanonicalMetricDataPoint[],
+  ): Promise<MetricRollupRow[]> {
+    const affected = new Set<number>();
+    for (const point of points) {
+      const neighbors = await this.immediateNeighbors(point);
+      for (const bucket of affectedRollupBuckets({
+        points: neighbors,
+        insertedPoint: point,
+      })) {
+        affected.add(bucket);
+      }
+    }
+    const authoritative = await this.pointsForBuckets(points[0]!, affected);
+    return buildMetricRollups({
+      points: authoritative,
+      affectedBuckets: affected,
+    });
+  }
+
+  async queryUsageEstimates(
+    query: MetricUsageEstimateQuery,
+  ): Promise<MetricUsageEstimate[]> {
+    if (!query.organizationId) {
+      throw new SecurityError(
+        "MetricDataPointClickHouseRepository.queryUsageEstimates",
+        "organizationId is required",
+      );
+    }
+    const client = query.tenantId
+      ? await this.resolveClient(query.tenantId)
+      : await this.resolveOrganizationClient(query.organizationId);
+    return await queryMetricUsageEstimates({ client, query });
   }
 
   private async immediateNeighbors(
@@ -319,172 +236,56 @@ export class MetricDataPointClickHouseRepository implements MetricDataPointRepos
       format: "JSONEachRow",
     });
     return (await result.json<RawMetricRow>()).map((row) =>
-      fromRaw(row, point.organizationId),
+      fromRaw({ row, organizationId: point.organizationId }),
     );
   }
 
+  /**
+   * Every point in the affected buckets, each preceded by the sample the fold
+   * differences it against. Buckets are fetched as their own narrow ranges
+   * rather than one span: a late point and a distant next sample would
+   * otherwise scan every partition between them only to discard the rows.
+   */
   private async pointsForBuckets(
     point: CanonicalMetricDataPoint,
     buckets: ReadonlySet<number>,
   ): Promise<CanonicalMetricDataPoint[]> {
-    const starts = [...buckets];
-    const from = Math.min(...starts);
-    const to = Math.max(...starts) + METRIC_ROLLUP_INTERVAL_MS;
+    const starts = [...buckets].sort((a, b) => a - b);
+    if (starts.length === 0) return [];
+    const params: Record<string, unknown> = {
+      tenantId: point.tenantId,
+      seriesId: point.seriesId,
+    };
+    const selects = starts.flatMap((start, index) => {
+      params[`from${index}`] = new Date(start);
+      params[`to${index}`] = new Date(start + METRIC_ROLLUP_INTERVAL_MS);
+      return [
+        `(SELECT ${RAW_SELECT}
+          FROM metric_data_points FINAL
+          WHERE TenantId = {tenantId:String} AND SeriesId = {seriesId:String}
+            AND TimeUnixMs < {from${index}:DateTime64(3)}
+          ORDER BY TimeUnixNano DESC, PointId DESC LIMIT 1)`,
+        `(SELECT ${RAW_SELECT}
+          FROM metric_data_points FINAL
+          WHERE TenantId = {tenantId:String} AND SeriesId = {seriesId:String}
+            AND TimeUnixMs >= {from${index}:DateTime64(3)}
+            AND TimeUnixMs < {to${index}:DateTime64(3)}
+          ORDER BY TimeUnixNano ASC, PointId ASC)`,
+      ];
+    });
     const client = await this.resolveClient(point.tenantId);
     const result = await client.query({
-      query: `
-        (SELECT ${RAW_SELECT}
-         FROM metric_data_points FINAL
-         WHERE TenantId = {tenantId:String} AND SeriesId = {seriesId:String}
-           AND TimeUnixMs < {from:DateTime64(3)}
-         ORDER BY TimeUnixNano DESC, PointId DESC LIMIT 1)
-        UNION ALL
-        (SELECT ${RAW_SELECT}
-         FROM metric_data_points FINAL
-         WHERE TenantId = {tenantId:String} AND SeriesId = {seriesId:String}
-           AND TimeUnixMs >= {from:DateTime64(3)} AND TimeUnixMs < {to:DateTime64(3)}
-         ORDER BY TimeUnixNano ASC, PointId ASC)
-      `,
-      query_params: {
-        tenantId: point.tenantId,
-        seriesId: point.seriesId,
-        from: new Date(from),
-        to: new Date(to),
-      },
+      query: selects.join("\n UNION ALL\n"),
+      query_params: params,
       format: "JSONEachRow",
     });
-    return (await result.json<RawMetricRow>()).map((row) =>
-      fromRaw(row, point.organizationId),
-    );
-  }
-
-  async recomputeAffectedRollups(
-    point: CanonicalMetricDataPoint,
-    retentionDays = PLATFORM_DEFAULT_RETENTION_DAYS,
-  ): Promise<void> {
-    // Projection queues are independent. Ensuring the point here makes the
-    // raw-before-derived invariant true even if this projection wins the race.
-    await this.ensureDataPoint(point, retentionDays);
-    const neighbors = await this.immediateNeighbors(point);
-    const affected = affectedRollupBuckets(neighbors, point);
-    const authoritative = await this.pointsForBuckets(point, affected);
-    const rows = buildMetricRollups(authoritative, affected);
-    if (rows.length === 0) return;
-    const client = await this.resolveClient(point.tenantId);
-    await client.insert({
-      table: "metric_time_rollups",
-      values: rows.map((row) => ({
-        TenantId: row.tenantId,
-        SeriesId: row.seriesId,
-        MetricName: row.metricName,
-        MetricUnit: row.metricUnit,
-        MetricKind: row.metricKind,
-        AggregationTemporality: row.aggregationTemporality,
-        IsMonotonic: row.isMonotonic,
-        BucketStart: new Date(row.bucketStartMs),
-        BucketEnd: new Date(row.bucketEndMs),
-        GaugeLast: row.gaugeLast,
-        Min: row.min,
-        Max: row.max,
-        Sum: row.sum,
-        Count: row.count,
-        ExplicitBounds: row.explicitBounds,
-        BucketCounts: row.bucketCounts,
-        ExponentialScale: row.exponentialScale,
-        ExponentialZeroThreshold: row.exponentialZeroThreshold,
-        ZeroCount: row.zeroCount,
-        PositiveOffset: row.positiveOffset,
-        PositiveBucketCounts: row.positiveBucketCounts,
-        NegativeOffset: row.negativeOffset,
-        NegativeBucketCounts: row.negativeBucketCounts,
-        ResetCount: row.resetCount,
-        GapCount: row.gapCount,
-        SourcePointCount: row.sourcePointCount,
-        UpdatedAt: new Date(row.updatedAt),
-        _retention_days: retentionDays,
-        _size_bytes: 0,
-      })),
-      format: "JSONEachRow",
-      clickhouse_settings: { async_insert: 1, wait_for_async_insert: 1 },
-    });
-  }
-
-  async queryUsageEstimates(
-    query: MetricUsageEstimateQuery,
-  ): Promise<MetricUsageEstimate[]> {
-    if (!query.organizationId) {
-      throw new SecurityError(
-        "MetricDataPointClickHouseRepository.queryUsageEstimates",
-        "organizationId is required",
-      );
+    // A bucket's predecessor may itself sit in an earlier affected bucket, so
+    // the ranges overlap by design; the fold needs each point exactly once.
+    const unique = new Map<string, CanonicalMetricDataPoint>();
+    for (const row of await result.json<RawMetricRow>()) {
+      const parsed = fromRaw({ row, organizationId: point.organizationId });
+      unique.set(parsed.pointId, parsed);
     }
-    const client = query.tenantId
-      ? await this.resolveClient(query.tenantId)
-      : await this.resolveOrganizationClient(query.organizationId);
-    const dimensions = {
-      organization: ["OrganizationId"],
-      project: ["OrganizationId", "TenantId"],
-      metric: ["OrganizationId", "TenantId", "MetricName"],
-      hour: ["OrganizationId", "TenantId", "MetricName", "AcceptedHour"],
-    }[query.groupBy];
-    const selectDimensions = dimensions.join(", ");
-    const identityWhere = [
-      "OrganizationId = {organizationId:String}",
-      "AcceptedAt < {to:DateTime64(3)}",
-      query.tenantId ? "TenantId = {tenantId:String}" : "",
-      query.metricName ? "MetricName = {metricName:String}" : "",
-    ]
-      .filter(Boolean)
-      .join(" AND ");
-    const result = await client.query({
-      query: `
-        SELECT
-          ${selectDimensions},
-          uniqExact(SeriesId) AS UniqueActiveSeries,
-          uniqExact(tuple(SeriesId, AcceptedHour)) AS ActiveSeriesHours,
-          uniqExact(PointId) AS AcceptedPoints,
-          sum(CanonicalSourceBytes) AS CanonicalRetainedBytes,
-          uniqExact(tuple(SeriesId, AcceptedHour)) AS ProjectedEventEquivalentUsage
-          FROM (
-          SELECT
-            PointId,
-            any(OrganizationId) AS OrganizationId,
-            any(TenantId) AS TenantId,
-            any(SeriesId) AS SeriesId,
-            any(MetricName) AS MetricName,
-            min(AcceptedAt) AS AcceptedAt,
-            toStartOfHour(min(AcceptedAt)) AS AcceptedHour,
-            any(CanonicalSourceBytes) AS CanonicalSourceBytes
-          FROM metric_usage_estimates
-          WHERE ${identityWhere}
-          GROUP BY PointId
-          HAVING min(AcceptedAt) >= {from:DateTime64(3)}
-        )
-        GROUP BY ${selectDimensions}
-        ORDER BY ${selectDimensions}
-      `,
-      query_params: {
-        organizationId: query.organizationId,
-        from: query.from,
-        to: query.to,
-        ...(query.tenantId ? { tenantId: query.tenantId } : {}),
-        ...(query.metricName ? { metricName: query.metricName } : {}),
-      },
-      format: "JSONEachRow",
-    });
-    const rows = await result.json<Record<string, string>>();
-    return rows.map((row) => ({
-      organizationId: row.OrganizationId!,
-      tenantId: row.TenantId ?? null,
-      metricName: row.MetricName ?? null,
-      acceptedHour: row.AcceptedHour ?? null,
-      uniqueActiveSeries: Number(row.UniqueActiveSeries ?? 0),
-      activeSeriesHours: Number(row.ActiveSeriesHours ?? 0),
-      acceptedPoints: Number(row.AcceptedPoints ?? 0),
-      canonicalRetainedBytes: Number(row.CanonicalRetainedBytes ?? 0),
-      projectedEventEquivalentUsage: Number(
-        row.ProjectedEventEquivalentUsage ?? 0,
-      ),
-    }));
+    return [...unique.values()];
   }
 }
