+++ langwatch/src/server/app-layer/metrics/repositories/metric-data-point.repository.ts
@@ -4,38 +4,42 @@ import type {
   MetricUsageEstimateQuery,
 } from "~/server/event-sourcing/pipelines/metric-processing/schemas/metricDataPoint";
 
+export interface MetricDataPointWrite {
+  point: CanonicalMetricDataPoint;
+  retentionDays?: number;
+}
+
+/** A replay chunk: many points for one tenant, written in one round trip. */
+export interface MetricDataPointBulkWrite {
+  points: CanonicalMetricDataPoint[];
+  retentionDays?: number;
+}
+
 export interface MetricDataPointRepository {
-  ensureDataPoint(
-    point: CanonicalMetricDataPoint,
-    retentionDays?: number,
-  ): Promise<void>;
-  upsertSeries(
-    point: CanonicalMetricDataPoint,
-    retentionDays?: number,
-  ): Promise<void>;
-  recomputeAffectedRollups(
-    point: CanonicalMetricDataPoint,
-    retentionDays?: number,
-  ): Promise<void>;
+  ensureDataPoint(args: MetricDataPointWrite): Promise<void>;
+  ensureDataPoints(args: MetricDataPointBulkWrite): Promise<void>;
+  upsertSeries(args: MetricDataPointWrite): Promise<void>;
+  upsertSeriesMany(args: MetricDataPointBulkWrite): Promise<void>;
+  recomputeAffectedRollups(args: MetricDataPointWrite): Promise<void>;
+  recomputeAffectedRollupsMany(args: MetricDataPointBulkWrite): Promise<void>;
   queryUsageEstimates(
     query: MetricUsageEstimateQuery,
   ): Promise<MetricUsageEstimate[]>;
 }
 
 export class NullMetricDataPointRepository implements MetricDataPointRepository {
-  async ensureDataPoint(
-    _point: CanonicalMetricDataPoint,
-    _retentionDays?: number,
-  ): Promise<void> {}
+  async ensureDataPoint(_args: MetricDataPointWrite): Promise<void> {}
 
-  async upsertSeries(
-    _point: CanonicalMetricDataPoint,
-    _retentionDays?: number,
-  ): Promise<void> {}
+  async ensureDataPoints(_args: MetricDataPointBulkWrite): Promise<void> {}
 
-  async recomputeAffectedRollups(
-    _point: CanonicalMetricDataPoint,
-    _retentionDays?: number,
+  async upsertSeries(_args: MetricDataPointWrite): Promise<void> {}
+
+  async upsertSeriesMany(_args: MetricDataPointBulkWrite): Promise<void> {}
+
+  async recomputeAffectedRollups(_args: MetricDataPointWrite): Promise<void> {}
+
+  async recomputeAffectedRollupsMany(
+    _args: MetricDataPointBulkWrite,
   ): Promise<void> {}
 
   async queryUsageEstimates(
