+++ langwatch/src/server/app-layer/metrics/repositories/metric-data-point.usage.ts
@@ -1,0 +1,97 @@
+import type { ClickHouseClient } from "@clickhouse/client";
+import type {
+  MetricUsageEstimate,
+  MetricUsageEstimateQuery,
+} from "~/server/event-sourcing/pipelines/metric-processing/schemas/metricDataPoint";
+
+const DIMENSIONS: Record<MetricUsageEstimateQuery["groupBy"], string[]> = {
+  organization: ["OrganizationId"],
+  project: ["OrganizationId", "TenantId"],
+  metric: ["OrganizationId", "TenantId", "MetricName"],
+  hour: ["OrganizationId", "TenantId", "MetricName", "AcceptedHour"],
+};
+
+/**
+ * Usage estimates over the shadow ledger.
+ *
+ * Tenant-isolation carve-out (see dev/docs/best_practices/clickhouse-queries.md,
+ * "Carve-out: organization-scoped billing ledgers"): `metric_usage_estimates`
+ * is the only table keyed by `(OrganizationId, TenantId, PointId)`, and its
+ * client is resolved from the organization — which is what selects the physical
+ * ClickHouse instance in the first place, so an organization can never read
+ * another's rows. `OrganizationId` leads the predicate because it leads the
+ * sort key; `TenantId` narrows it whenever the caller supplies one.
+ */
+export async function queryMetricUsageEstimates({
+  client,
+  query,
+}: {
+  client: ClickHouseClient;
+  query: MetricUsageEstimateQuery;
+}): Promise<MetricUsageEstimate[]> {
+  const dimensions = DIMENSIONS[query.groupBy];
+  const selectDimensions = dimensions.join(", ");
+  const identityWhere = [
+    "OrganizationId = {organizationId:String}",
+    // Upper bound only. A point is billed at its FIRST acceptance, so the
+    // min() below must see every earlier retry of the same PointId — a
+    // `AcceptedAt >= {from}` predicate here would hide the original row and
+    // re-bill a point whose first acceptance predates the window. The 13-month
+    // TTL bounds what the missing lower bound can scan.
+    "AcceptedAt < {to:DateTime64(3)}",
+    query.tenantId ? "TenantId = {tenantId:String}" : "",
+    query.metricName ? "MetricName = {metricName:String}" : "",
+  ]
+    .filter(Boolean)
+    .join(" AND ");
+
+  const result = await client.query({
+    query: `
+      SELECT
+        ${selectDimensions},
+        uniqExact(SeriesId) AS UniqueActiveSeries,
+        uniqExact(tuple(SeriesId, AcceptedHour)) AS ActiveSeriesHours,
+        uniqExact(PointId) AS AcceptedPoints,
+        sum(CanonicalSourceBytes) AS CanonicalRetainedBytes,
+        uniqExact(tuple(SeriesId, AcceptedHour)) AS ProjectedEventEquivalentUsage
+        FROM (
+        SELECT
+          PointId,
+          any(OrganizationId) AS OrganizationId,
+          any(TenantId) AS TenantId,
+          any(SeriesId) AS SeriesId,
+          any(MetricName) AS MetricName,
+          min(AcceptedAt) AS AcceptedAt,
+          toStartOfHour(min(AcceptedAt)) AS AcceptedHour,
+          any(CanonicalSourceBytes) AS CanonicalSourceBytes
+        FROM metric_usage_estimates
+        WHERE ${identityWhere}
+        GROUP BY PointId
+        HAVING min(AcceptedAt) >= {from:DateTime64(3)}
+      )
+      GROUP BY ${selectDimensions}
+      ORDER BY ${selectDimensions}
+    `,
+    query_params: {
+      organizationId: query.organizationId,
+      from: query.from,
+      to: query.to,
+      ...(query.tenantId ? { tenantId: query.tenantId } : {}),
+      ...(query.metricName ? { metricName: query.metricName } : {}),
+    },
+    format: "JSONEachRow",
+  });
+
+  const rows = await result.json<Record<string, string>>();
+  return rows.map((row) => ({
+    organizationId: row.OrganizationId!,
+    tenantId: row.TenantId ?? null,
+    metricName: row.MetricName ?? null,
+    acceptedHour: row.AcceptedHour ?? null,
+    uniqueActiveSeries: Number(row.UniqueActiveSeries ?? 0),
+    activeSeriesHours: Number(row.ActiveSeriesHours ?? 0),
+    acceptedPoints: Number(row.AcceptedPoints ?? 0),
+    canonicalRetainedBytes: Number(row.CanonicalRetainedBytes ?? 0),
+    projectedEventEquivalentUsage: Number(row.ProjectedEventEquivalentUsage ?? 0),
+  }));
+}
