+++ langwatch/src/server/event-sourcing/pipelines/metric-processing/__tests__/fixtures/metric-point.fixtures.ts
@@ -1,0 +1,54 @@
+import type { CanonicalMetricDataPoint } from "../../schemas/metricDataPoint";
+
+/** A canonical point with only the fields a rollup test cares about set. */
+export function point(
+  overrides: Partial<CanonicalMetricDataPoint> & { timeUnixMs: number },
+): CanonicalMetricDataPoint {
+  return {
+    tenantId: "project-1",
+    organizationId: "organization-1",
+    pointId: String(overrides.timeUnixMs).padStart(64, "0"),
+    seriesId: "a".repeat(64),
+    resourceSchemaUrl: "",
+    resourceAttributesJson: "[]",
+    resourceAttributeKeys: [],
+    scopeSchemaUrl: "",
+    scopeName: "scope",
+    scopeVersion: "",
+    scopeAttributesJson: "[]",
+    scopeAttributeKeys: [],
+    metricName: "metric",
+    metricDescription: "",
+    metricUnit: "1",
+    metricKind: "gauge",
+    aggregationTemporality: "unspecified",
+    isMonotonic: null,
+    pointAttributesJson: "[]",
+    pointAttributeKeys: [],
+    startTimeUnixNano: "1",
+    timeUnixNano: String(BigInt(overrides.timeUnixMs) * 1_000_000n),
+    flags: 0,
+    valueType: "double",
+    valueInt: null,
+    valueDouble: null,
+    count: null,
+    sum: null,
+    min: null,
+    max: null,
+    explicitBounds: [],
+    bucketCounts: [],
+    exponentialScale: null,
+    exponentialZeroThreshold: null,
+    zeroCount: null,
+    positiveOffset: null,
+    positiveBucketCounts: [],
+    negativeOffset: null,
+    negativeBucketCounts: [],
+    summaryQuantilesJson: "[]",
+    canonicalPayload: "{}",
+    canonicalSizeBytes: 2,
+    occurredAt: overrides.timeUnixMs,
+    acceptedAt: 1_800_000_000_000,
+    ...overrides,
+  };
+}
