+++ langwatch/src/server/event-sourcing/pipelines/trace-processing/projections/traceSummary.foldProjection.ts
@@ -12,7 +12,10 @@ import {
   type FoldEventHandlers,
 } from "~/server/event-sourcing/projections/abstractFoldProjection";
 import type { FoldProjectionStore } from "~/server/event-sourcing/projections/foldProjection.types";
-import { TRACE_SUMMARY_PROJECTION_VERSION_LATEST } from "../schemas/constants";
+import {
+  METRIC_EXEMPLAR_CORRELATION_COUNT_ATTRIBUTE,
+  TRACE_SUMMARY_PROJECTION_VERSION_LATEST,
+} from "../schemas/constants";
 import type {
   LogRecordReceivedEvent,
   MetricDataPointCorrelatedEvent,
@@ -551,13 +554,16 @@ export class TraceSummaryFoldProjection
           : Math.min(timeToFirstTokenMs, ttftMs);
     }
 
+    // Counts exemplar correlations, not metric data points: the canonical
+    // datapoint stream is a separate pipeline this fold never sees, so it
+    // cannot know how many points a trace's metrics produced.
     const mergedAttributes = { ...state.attributes };
-    const metricCount = parseInt(
-      mergedAttributes["langwatch.reserved.metric_record_count"] ?? "0",
+    const correlationCount = parseInt(
+      mergedAttributes[METRIC_EXEMPLAR_CORRELATION_COUNT_ATTRIBUTE] ?? "0",
       10,
     );
-    mergedAttributes["langwatch.reserved.metric_record_count"] = String(
-      metricCount + 1,
+    mergedAttributes[METRIC_EXEMPLAR_CORRELATION_COUNT_ATTRIBUTE] = String(
+      correlationCount + 1,
     );
 
     return {
